+++ demos/middleware/pages/index.js
@@ -26,6 +26,9 @@ export default function Home() {
           <Link href="/shows/rewrite-external">Rewrite to external URL</Link>
         </p>
         <p>
+          <Link href="/shows/static/3">Add header to static page</Link>
+        </p>
+        <p>
           <Link href="/cookies" prefetch={false}>
             Cookie API
           </Link>