+++ plugin/src/templates/ipx.ts
@@ -3,10 +3,11 @@ import { Handler } from '@netlify/functions'
 import { createIPXHandler } from '@netlify/ipx'
 
 // @ts-ignore Injected at build time
-import { basePath, domains } from './imageconfig.json'
+import { basePath, domains, remotePatterns } from './imageconfig.json'
 
 export const handler: Handler = createIPXHandler({
   basePath,
   domains,
+  remotePatterns,
 }) as Handler
 /* eslint-enable node/no-missing-import, import/no-unresolved, @typescript-eslint/ban-ts-comment  */